refactor: consolidate AgentCategory to single source of truth
- Create src/shared/schemas/agentCategory.ts as canonical source
- Move proposal field schemas to src/shared/schemas/proposalFields.ts
- Update all imports to use @shared/schemas directly
- Delete duplicate src/common/constants/agentTypes.js
- Fix TypeScript TS2702 errors using `typeof AgentCategory.X` in type positions
- Update HTML webview imports to use shared module path
