+++ src/agent/core/AgentConfig.ts
@@ -2,58 +2,14 @@
 import { z } from 'zod';
 
 // Local imports - agent components
-import { AgentCategory } from './AgentDataclass';
+import { AgentCategory, AgentCategorySchema } from './AgentDataclass';
 import { DEFAULT_TOOL_CONFIG, ToolConfigSchema } from './ToolConfig';
 
-/**
- * Base proposal fields shared by both workflow and tool-use agent proposals.
- * Contains only the common fields that all proposal types need.
- */
-export const BaseProposalFieldsSchema = z.object({
-  agent: z.string().describe('Name of the agent to execute'),
-  model: z.string().describe('Model to use for agent execution'),
-  instruction: z.string().describe('Instruction for the agent'),
-});
-
-/**
- * File path fields for workflow agents only.
- * Tool-use agents access files through their own tools instead.
- * Used internally to compose WorkflowSpecificFieldsSchema.
- */
-const FileFieldsSchema = z.object({
-  inputFile: z.string().describe('Path to the primary input file'),
-  inputFiles: z.array(z.string()).describe('Additional input file paths'),
-  referenceFile: z
-    .string()
-    .nullable()
-    .describe('Reference file path for additional context'),
-  referenceFiles: z
-    .array(z.string())
-    .describe('Additional reference file paths'),
-  auxiliaryFile: z
-    .string()
-    .nullable()
-    .describe('Auxiliary file path for supplementary content'),
-  auxiliaryFiles: z
-    .array(z.string())
-    .describe('Additional auxiliary file paths'),
-  mediaFile: z
-    .string()
-    .nullable()
-    .describe('Media file path for images/figures'),
-  mediaFiles: z.array(z.string()).describe('Additional media file paths'),
-  outputFiles: z.array(z.string()).describe('Desired output file paths'),
-});
-
-/**
- * Workflow-specific fields: file fields + multiple outputs flag.
- * Only workflow agents (document processing) use these fields.
- */
-export const WorkflowSpecificFieldsSchema = FileFieldsSchema.extend({
-  useMultipleOutputs: z
-    .boolean()
-    .describe('Enable multiple outputs mode for agents that support it'),
-});
+// Re-export proposal schemas from shared (single source of truth)
+export {
+  BaseProposalFieldsSchema,
+  WorkflowSpecificFieldsSchema,
+} from '@shared/schemas/proposalFields';
 
 /** Zod schema for validating AgentConfig objects */
 const stringArrayField = () => z.array(z.string()).prefault([]);
@@ -79,7 +35,7 @@ const AgentConfigFieldsSchema = z.object({
   outputFiles: stringArrayField(),
 
   // AgentConfig-specific fields
-  agentCategory: z.enum(AgentCategory).prefault(AgentCategory.Workflow),
+  agentCategory: AgentCategorySchema.prefault(AgentCategory.Workflow),
   editedFile: z.string().nullable().prefault(null),
   editedFiles: stringArrayField(),
