--- src/common/constants/agentTypes.js
@@ -1,11 +1,0 @@
-// Agent category constants - single source of truth for agent/session categories
-// Used across webview, historyView, and other modules
-
-/**
- * Agent category types.
- * These values must match the AgentCategory enum in AgentDataclass.ts.
- */
-export const AGENT_CATEGORY = {
-  WORKFLOW: 'workflow',
-  TOOL_USE: 'toolUse',
-};
