+++ src/historyView/index.html
@@ -40,7 +40,7 @@
           "@common/BaseWebviewMessageHandler.js": "${baseWebviewMessageHandlerUri}",
           "@common/BaseDomHandler.js": "${baseDomHandlerUri}",
           "@common/ToggleStateStore.js": "${toggleStateStoreUri}",
-          "@common/constants/agentTypes.js": "${agentCategoriesUri}",
+          "@shared/schemas/agentCategory.js": "${agentCategoriesUri}",
           "he": "https://esm.sh/he@1.2.0"
         }
       }
