+++ src/progressView/state/ProgressViewState.ts
@@ -2,7 +2,7 @@
 import { z } from 'zod';
 
 // Local imports - agent metadata
-import { AgentCategory } from '@agent/core/AgentDataclass';
+import { AgentCategorySchema } from '@shared/schemas';
 import { isAgentCategoryFilter } from '@agent/types/AgentStreamTypes';
 // Type imports
 import {
@@ -45,7 +45,7 @@ import { TodoItemSchema, type TodoItem } from '@shared/schemas';
  * Used to display UI indicators before TaskState is fully populated.
  */
 export const StreamHintsSchema = z.object({
-  agentCategory: z.enum(AgentCategory).optional(),
+  agentCategory: AgentCategorySchema.optional(),
   isRemote: z.boolean().optional(),
   hasMultipleOutputs: z.boolean().optional(),
 });
