+++ src/shared/schemas/index.ts
@@ -1,8 +1,10 @@
+export * from './agentCategory';
 export * from './errors';
 export * from './identifiers';
 export * from './log';
 export * from './output';
 export * from './prompts';
+export * from './proposalFields';
 export * from './stream';
 export * from './taskGroup';
 export * from './todo';
