+++ src/webview/index.html
@@ -49,7 +49,7 @@
           "@common/iconConstants.js": "${iconConstantsUri}",
           "@common/dropdownUtils.js": "${dropdownUtilsUri}",
           "@common/BaseWebviewMessageHandler.js": "${baseWebviewMessageHandlerUri}",
-          "@common/constants/agentTypes.js": "${agentCategoriesUri}",
+          "@shared/schemas/agentCategory.js": "${agentCategoriesUri}",
           "sortablejs": "https://unpkg.com/sortablejs@1.15.6/modular/sortable.esm.js",
           "nanoid": "https://unpkg.com/nanoid@5.1.6/index.browser.js"
         }
