+++ src/webview/modules/constants.js
@@ -1,6 +1,6 @@
 // Local imports - webview
 import { MAIN_VIEW_COMMANDS } from '@common/webview/commands.js';
-import { AGENT_CATEGORY } from '@common/constants/agentTypes.js';
+import { AGENT_CATEGORY } from '@shared/schemas/agentCategory.js';
 
 // Basic file types
 export const FILE_TYPES = [
